requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
outcome==1.2.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
rsa==4.9
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.13.0